Refactored the main file to use argparse to dispatch the commands
This allows us to do a couple of things differently (and in my opinion better). First, it eliminates the ever-growing chain of if/else blocks that were previously used to dispatch commands and replaces it with an argument parser that parses the input and calls the appropriate methods. It also allows us to group similar commands to reduce the number of top level commands the user needs to use. The user can also pass the optional arguments in any order they wish.

For example, under this implementation we can get scores (which are logically related to games) by adding a flag to the Game command  as follows:
Game -s -y 2020 -w 10
Rather than 'Scores 2020 10', which is a separate command

We can also provide the user with help methods, either at the top level (ex. -h), or at the command level (command -h).

This does change the login method from NFLapp <username> <password> to NFLapp Login -u <username> -p <password>.